requirements.txt
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
easyfsl==1.5.0
filelock==3.9.0
fonttools==4.46.0
fsspec==2023.4.0
idna==3.4
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.8.2
mpmath==1.3.0
natsort==8.4.0
networkx==3.0
numpy==1.24.1
packaging==23.2
pandas==2.1.3
Pillow==9.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sympy==1.12
torchinfo==1.8.0
tqdm==4.66.1
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
zipp==3.17.0

--extra-index-url https://download.pytorch.org/whl/cu121
    torch
    torchvision
    torchaudio
